refactor: restructure project using vertical slice architecture
