+++ DiagramPersistenceService/src/Core/Classes/UmlClass.cs
@@ -1,6 +1,6 @@
 using OpenDDD.Domain.Model.Base;
 
-namespace Core.Model;
+namespace Core.Classes;
 
 /// <summary>
 /// Represents a UML class in a diagram.
