+++ DiagramPersistenceService/src/Core/Classes/UmlMethod.cs
@@ -1,6 +1,6 @@
 using OpenDDD.Domain.Model.Base;
 
-namespace Core.Model;
+namespace Core.Classes;
 
 /// <summary>
 /// Represents a method of a UML class.
