+++ DiagramPersistenceService/src/Core/Classes/UmlMethodParameter.cs
@@ -1,6 +1,6 @@
 using OpenDDD.Domain.Model.Base;
 
-namespace Core.Model;
+namespace Core.Classes;
 
 /// <summary>
 /// Represents a parameter of a UML method.
