+++ DiagramPersistenceService/src/Core/Core.csproj
@@ -1,18 +1,10 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-
   <PropertyGroup>
     <TargetFramework>net9.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
-
   <ItemGroup>
     <PackageReference Include="OpenDDD.NET" Version="3.0.0-beta.2" />
   </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Model/" />
-    <Folder Include="Services/" />
-  </ItemGroup>
-
 </Project>
